+++ source/win-main/GlobalSearch.vue
@@ -112,7 +112,7 @@
             v-bind:key="idx2"
             class="result-line"
             v-bind:class="{'active': idx==activeFileIdx && idx2==activeLineIdx}"
-            v-on:contextmenu.stop.prevent="fileContextMenu($event, result.file.path, singleRes.line)"
+            v-on:contextmenu.stop.prevent="fileContextMenu($event, result.file.path, singleRes.line, singleRes.restext)"
             v-on:mousedown.stop.prevent="onResultClick($event, idx, idx2, result.file.path, singleRes.line)"
           >
             <!-- NOTE how we have to increase the line number from zero-based to 1-based -->
@@ -181,6 +181,12 @@ function getContextMenu (): AnyMenuItem[] {
     {
       label: trans('Open in new tab'),
       id: 'new-tab',
+      type: 'normal',
+      enabled: true
+    },
+    {
+      label: trans('Copy'),
+      id: 'copy',
       type: 'normal',
       enabled: true
     }
@@ -471,12 +477,15 @@ function toggleIndividualResults (): void {
   }
 }
 
-function fileContextMenu (event: MouseEvent, filePath: string, lineNumber: number): void {
+function fileContextMenu (event: MouseEvent, filePath: string, lineNumber: number, restext: string): void {
   const point = { x: event.clientX, y: event.clientY }
   showPopupMenu(point, getContextMenu(), (clickedID: string) => {
     switch (clickedID) {
       case 'new-tab':
         jumpToLine(filePath, lineNumber, true)
+        break
+      case 'copy':
+        navigator.clipboard.writeText(restext).catch(err => console.error(err))
         break
     }
   })